Test agents (#2)
* update

* update Agents

* change Position variables

int to double for better precision

+++ src/elements/Point.java
@@ -2,13 +2,13 @@ package elements;
 
 public class Point {
 
-    int x, y;
+    double x, y;
 
     /**
      * @param x - position in xx
      * @param y - position in yy
      */
-    public Point(int x, int y) {
+    public Point(double x, double y) {
         this.x = x;
         this.y = y;
     }
@@ -18,7 +18,7 @@ public class Point {
      * @return - position in xx
      */
     
-    public int getX() {
+    public double getX() {
         return x;
     }
     
@@ -27,7 +27,7 @@ public class Point {
      * @return - position in yy
      */
 
-    public int getY() {
+    public double getY() {
         return y;
     }
